+++ client/dashboard/CHANGELOG.md
@@ -1,5 +1,28 @@
 # dashboard
 
+## 0.26.3
+
+### Patch Changes
+
+- 12e825c: Add hide/show toggle for environment variable inputs
+
+## 0.26.2
+
+### Patch Changes
+
+- 81be736: Updates dashboard to only use telemetry API
+- Updated dependencies [f2fa135]
+  - @gram-ai/elements@1.19.1
+
+## 0.26.1
+
+### Patch Changes
+
+- Updated dependencies [856576b]
+- Updated dependencies [a1231be]
+- Updated dependencies [748c52e]
+  - @gram-ai/elements@1.19.0
+
 ## 0.26.0
 
 ### Minor Changes